Split DefaultService and change Default.js accordingly
