+++ controllers/Default.js
@@ -2,9 +2,13 @@
 
 var utils = require('../utils/writer.js');
 var Default = require('../service/DefaultService');
+var Catalog = require('../service/CatalogService');
+var Goals = require('../service/GoalsService');
+var Planner = require('../service/PlannerService');
+var Reservations = require('../service/ReservationsService');
 
 module.exports.cancelReservation = function cancelReservation (req, res, next, username, day, time) {
-  Default.cancelReservation(username, day, time)
+  Reservations.cancelReservation(username, day, time)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -14,7 +18,7 @@ module.exports.cancelReservation = function cancelReservation (req, res, next, u
 };
 
 module.exports.checkGoalsFromInfo = function checkGoalsFromInfo (req, res, next, username, currentBodyWeight) {
-  Default.checkGoalsFromInfo(username, currentBodyWeight)
+  Goals.checkGoalsFromInfo(username, currentBodyWeight)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -25,7 +29,7 @@ module.exports.checkGoalsFromInfo = function checkGoalsFromInfo (req, res, next,
 };
 
 module.exports.checkGoalsFromProgress = function checkGoalsFromProgress (req, res, next, username, day) {
-  Default.checkGoalsFromProgress(username, day)
+  Goals.checkGoalsFromProgress(username, day)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -35,7 +39,7 @@ module.exports.checkGoalsFromProgress = function checkGoalsFromProgress (req, re
 };
 
 module.exports.createCustomExercise = function createCustomExercise (req, res, next, body, username) {
-  Default.createCustomExercise(body, username)
+  Catalog.createCustomExercise(body, username)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -45,7 +49,7 @@ module.exports.createCustomExercise = function createCustomExercise (req, res, n
 };
 
 module.exports.getAvailableReservations = function getAvailableReservations (req, res, next, username, day) {
-  Default.getAvailableReservations(username, day)
+  Reservations.getAvailableReservations(username, day)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -55,7 +59,7 @@ module.exports.getAvailableReservations = function getAvailableReservations (req
 };
 
 module.exports.getDayofPlanner = function getDayofPlanner (req, res, next, username, day) {
-  Default.getDayofPlanner(username, day)
+  Planner.getDayofPlanner(username, day)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -65,7 +69,7 @@ module.exports.getDayofPlanner = function getDayofPlanner (req, res, next, usern
 };
 
 module.exports.getDropDownMenuList = function getDropDownMenuList (req, res, next, username) {
-  Default.getDropDownMenuList(username)
+  Catalog.getDropDownMenuList(username)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -75,7 +79,7 @@ module.exports.getDropDownMenuList = function getDropDownMenuList (req, res, nex
 };
 
 module.exports.getExerciseCatalog = function getExerciseCatalog (req, res, next, username) {
-  Default.getExerciseCatalog(username)
+  Catalog.getExerciseCatalog(username)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -85,7 +89,7 @@ module.exports.getExerciseCatalog = function getExerciseCatalog (req, res, next,
 };
 
 module.exports.getExerciseProgress = function getExerciseProgress (req, res, next, username, exerciseName) {
-  Default.getExerciseProgress(username, exerciseName)
+  Planner.getExerciseProgress(username, exerciseName)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -95,7 +99,7 @@ module.exports.getExerciseProgress = function getExerciseProgress (req, res, nex
 };
 
 module.exports.getMyReservations = function getMyReservations (req, res, next, username) {
-  Default.getMyReservations(username)
+  Reservations.getMyReservations(username)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -116,7 +120,7 @@ module.exports.getPersonalInfo = function getPersonalInfo (req, res, next, usern
 
 module.exports.makeReservation = function makeReservation (req, res, next, body, username) {
   
-  Default.makeReservation(body, username)
+  Reservations.makeReservation(body, username)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
@@ -126,7 +130,7 @@ module.exports.makeReservation = function makeReservation (req, res, next, body,
 };
 
 module.exports.updateExerciseProgress = function updateExerciseProgress (req, res, next, day, name, weight, reps, username) {
-  Default.updateExerciseProgress(day, name, weight, reps, username)
+  Planner.updateExerciseProgress(day, name, weight, reps, username)
     .then(function (response) {
       utils.writeJson(res, response, response.code);
     })
